clean up utils includes
