+++ config_utilities/src/asl_formatter.cpp
@@ -250,7 +250,7 @@ std::string AslFormatter::formatField(const FieldInfo& info, size_t indent) cons
   const auto& settings = Settings::instance().printing;
 
   // field is the stringified value, The header is the field name.
-  std::string field = dataToString(info.value, settings.reformat_floats);
+  std::string field = yamlToString(info.value, settings.reformat_floats);
   if (info.isDefault() && Settings::instance().printing.show_defaults) {
     field += " (default)";
   }
